Add files via upload
Here we are trying to detect the area of interest.
The lanes we are trying to detect a parallel lines. However in the picture we see that at point they will converge. Thus we see a polygon(a traingle) being formed.
So we will plot the three points needed to get the desired area of interest. Removing(Blackening) the rest of the region.